fix: cast to uint64_t for correct value extraction in response handling

+++ esphome/components/sofarsolar_inverter/sofarsolar_inverter.cpp
@@ -284,7 +284,7 @@ namespace esphome {
                 case 4:
                     return (response.data()[3] << 24) | (response.data()[4] << 16) | (response.data()[5] << 8) | response.data()[6]; // 4 bytes for start address
                 case 8:
-                    return (response.data()[3] << 56) | (response.data()[4] << 48) | (response.data()[5] << 40) | (response.data()[6] << 32) |
+                    return (uint64_t) (response.data()[3] << 56) | (response.data()[4] << 48) | (response.data()[5] << 40) | (response.data()[6] << 32) |
                            (response.data()[7] << 24) | (response.data()[8] << 16) | (response.data()[9] << 8) | response.data()[10]; // 8 bytes for start address
                 default:
                     ESP_LOGE(TAG, "Invalid response size for extracting value");
